optimize directory structure
