+++ cmd/migrate/migrate.go
@@ -1,0 +1,1 @@
+package migrate
