+++ global/global.go
@@ -1,0 +1,3 @@
+package global
+
+var ServerConfigInstance *serverConfig
