+++ middleware/auth.go
@@ -1,0 +1,1 @@
+package middleware
