+++ models/request/login_request.go
@@ -1,4 +1,4 @@
-package validator
+package request
 
 type LoginRequest struct {
 	Username string `json:"username" binding:"required"`
