+++ models/request/test_reqeust.go
@@ -1,0 +1,6 @@
+package request
+
+type TestLoginRequest struct {
+	Username string `json:"username" validate:"required"`
+	Password string `json:"password" validate:"required"`
+}
