+++ types/constants/status_code.go
@@ -1,0 +1,1 @@
+package constants
