+++ types/enums/user_role.go
@@ -1,0 +1,1 @@
+package enums
