+++ types/errors/custom_errors.go
@@ -1,0 +1,1 @@
+package errors
